Closes #15628 by making var names in generated Cython code unique
Previously, the following call to `unfuncify` would result in non-compiling
Cython code due to a variable clash:

    >>> unfuncify((x, y), x + y, backend='Cython')

This was because the resulting Cython code looked something like:

    def autofunc_c(np.ndarray[np.double_t, ndim=1] _x, np.ndarray[np.double_t, ndim=1] _y):
        cdef int _m = _y.shape[0]
        cdef np.ndarray[np.double_t, ndim=1] _y = np.empty((_m))

where the argument `_y` is declared in the function body even though it is
already passed as an argument. This commit leverages `doprint` within the C code
generator to make sure that generated variable names are unique.