Update includes my attempt to show the contour to the user.
It is working for theta in (0, pi) but giving strange curve for pi/4. Also, it would be good if y-axis has like "7i" instead of "7" without the loss of generality and if that temporary "radius=1" can be replaced by radius=oo . For poles in denominator there's no problem but how do I calculate points where the numerator is oo since solve gives an empty list. One case can be when a number is not a pole due to cancellation which I would find a way but for now please help me.