allow control of Matrix.diag from dense.diag
dense.diag is the public facing diag routine that
never got merged with common.diag (accessible through Matrix.diag)
and each routine has different semantics when it comes to
unpacking and treatment of arguments that are lists.

Matrix.diag unpacked singleton sequences and did not convert other
lists to matrices

dense.diag converted all sequences to matrices and hence did not
unpack args

The behavior of dense.diag -- the publicly facing routine imported
with 'from sympy import *' is unchanged.

The behavior of common.diag a.k.a Matrix.diag has been modified to
treat lists as matrices rather than to try place list-literals
into a matrix (which creates a non-functional matrix):

>>> Matrix.diag([1,2],[3])
Matrix([
[[1, 2],   0],
[     0, [3]]])

The "wandering diagonal" behavior is retained so a
block-diagonal can be created from non-square matrices.
If the elements always start filling on the main diagonal,
blocks will end up overlapping. So the folling is correct:

>>> diag([1,2,3],4)
Matrix([
[1, 0],
[2, 0],
[3, 0],
[0, 4]])

This is not a block-diagonal since the "4" block overlaps
with the first column.

>>> diag([1,2,3],4)
Matrix([
[1, 0],
[2, 4],
[3, 0]])