Change return type of sort_key function
The return type for this function defined in `partitions.py` was list.
The error being raised was that the type list was not hashable.
But tuple is hashable, so according to the suggestions, I changed the return value into a tuple.
I wrote a test in `test_partitions.py` and it passed when I ran it with `bin/test`.