Add comments that MatAdd and MatMul probably shouldn't be subclassing from Add and Mul
There are some test failures right now if you remove them, at least in the
printers because they are using the Add and Mul printers, possibly other
places as well. For now we can leave it be but leave a comment that we might
want to remove them if they cause problems.