Remove Python 3.4 support
Python 3.4 has reached its [end of
life](https://devguide.python.org/#status-of-python-branches), so by our
[Python version support
policy](https://github.com/sympy/sympy/wiki/Python-version-support-policy) we
are no longer supporting it.

SymPy 1.4 will be the last version to support Python 3.4. Now that the 1.4
release branch has been made, we no longer need to support it on master.

It is unlikely this will buy us anything in the way of Python features, since
we still are supporting Python 2.7. However, it does mean fewer tests on
Travis.