Remove leading term created by rounding error
In a long division step of polynomials, the leading coefficient  a
of the dividend  f  is divided by the leading coefficient  b  of the
divisor g and the remainder  r = f - (a/b)*x**d*g  is constructed.
This will cancel the leading term of  f  if  a - (a/b)*b  is zero.
However, in an inexact domain like RR there may remain a small nonzero
residue. This patch will replace it by zero.

Define is_Exact and get_exact for FractionField.
If the coefficients of a composite domain are not exact, the domain
should not be considered exact. The definitions should probably be
added to PolynomialRing as well.

Add check for composite inexact domains in `roots`. The algorithm was
coded assuming that only numeric domains would be inexact.

Add tests for issues 2884, 7864 and 11542.