Add a check to mailmap_update for names that are email addresses
We can also check for invalid email addresses (no @), but I have left this out
for now, as the name check seems sufficient to find people who get the name
and email backwards. There is at least one person in the git history with an
empty email, which would have to be replaced with a stand in email like
none@example.com.