log(-oo) -> oo + I*pi
#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234". See
https://github.com/blog/1506-closing-issues-via-pull-requests . Please also
write a comment on that issue linking back to this pull request once it is
open. -->
close #16737

#### Brief description of what is fixed or changed
log(-oo) now gives oo + I*pi instead of oo

This is consistent with log(float(-inf')) in master.

#### Other comments

Also removed a redundant handling of zz and included an `if` in the `if-else` block.
#### Release Notes

<!-- Write the release notes for this release below. See
https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more information
on how to write release notes. The bot will check your release notes
automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* functions
    * log(-oo) now returns complex `oo + I*pi`
<!-- END RELEASE NOTES -->