refactor number.__rel__ code
An important change here is that when the Rational is
recast for comparison with the Float, the Float's
precision is used.

This always affects rationals that are inexact in
base-2 representation (i.e. don't have a denominator that
is a power of 2)

>>> round_trips = lambda n: Rational(Float(n)) == n

>>> round_trip(S('1/3'))
False

It only affects exact rationals with
a numerator with more bits than the requested precision.

>>> round_trips(S('1/2'))
True

>>> odd = lambda n: prod([2*i+1 for i in range(1, n)])
>>> round_trips(odd(15)/S(2))
True
>>> round_trips(odd(16)/S(2))
False

The moral of the story is that Float, like Integer, will
truncate Rationals:

>>> odd(10)
654729075
>>> Rational(Float(_/S(2)))
654729075/2

>>> odd(20)
319830986772877770815625
>>> Rational(Float(_/S(2)))
159915493386438876069888