reduce multiple sympifications of iterables
When a known iterable is encountered, we have to check for
rational and strings needing conovert_xor=False. But our
wrapper is also going to sympify everything. So to try
reduce the amount of sympification needed we only do
the necessary sympification during sanitization,
and let the wrapper take care of the rest.

floats and any non-Python values that became Float
will be converted at the same time if the rational
flag is True.