Add optimization of numpy code involving matrix inverses
Adds an `optimization` option to the code printer base class that
executes any `codegen.rewriting` optimizations the printer object might
have. Also adds an optimization to the numpy printer that uses the
assumption system to rewrite the multiplication of a matrix and a vector
as the faster `np.linalg.solve` when appropriate.

For example:

    >>> from sympy.assumptions import assuming, Q
    >>> from sympy.matrices import MatrixSymbol
    >>> from sympy.printing.pycode import NumPyPrinter
    >>> A = MatrixSymbol('A', 2, 2)
    >>> x = MatrixSymbol('x', 2, 1)
    >>> p = NumPyPrinter(settings={'optimize' : True})
    >>> print(p.doprint(A**(-1) * x))
    (numpy.linalg.inv(A)).dot(x)
    >>> with assuming(Q.fullrank(A)):
    ...     p.doprint(A**(-1) * x)
    numpy.linalg.solve(A, x)