Move to representing solve as an AST node
Removes the `optimize` option introduced to the code printer in the
previous commit, instead instroducing the inverse optimization as a
top-level variable in `codegen.rewriting` that has to be applied by the
user. Also abstracts the solving operation as a `MatrixSolve` AST node
that can be handled by each code printer uniquely.