matrices: cleaned up A**n cases
Cleaned up some not well defined cases in MatrixArithmetic.__pow__()
which would sometimes give erroneous results such as raising an error
on a matrix raised to a valid positive integer symbol. Syncronized
this with MatPow.doit() so they are consistent. Added tests for
zero-determinant matrices to negative power as well as verifying
A**n == MatPow(A, n).doit().