Updated value for hbar to reflect current standard
In June 2019, the recommended value for the reduced Planck's
constant was set to a fixed quantity. Updating this quantity
necessitated updating the value of G, the Newtonian gravitational
constant, and kB, the Boltzmann constant, for internal consistency.
Furthermore, the unit consistency tests were updated. The values
for the tests were based on the 2018 CODATA values.