exp.as_leading_term() returns wrong results (Issue 4615)
as_leading_term does not work properly on exponential function.
As noted in the issue 4615, `exp((x+1)/x**2).as_leading_term(x)`
returns `exp(x**(-2))`, which is not correct.
Indeed when the argument of `exp` tends to `oo`, then it is not
correct doing the leading term of the arg and then apply the exponential.

Also SymPy in the case
`exp((2*x + 3) / (x+1)).as_leading_term(x)`
returns `S(1)`, which is not correct here, since the argument
is closed to `3`.

In `test_exponential.py` the following tests are added
```
    assert exp((x + 1) / x**2).as_leading_term(x) == exp((x + 1) / x**2)
    assert exp((2*x + 3) / (x+1)).as_leading_term(x) == exp(3)
```