Wrong results for as_leading_term()
The as_leading_term() for both factorial and gamma functions
return wrong result whenever the arguments tend to infinity.
Indeed gamma(2 + 1/n).as_leading_term() and factorial(2 + 1/n).as_leading_term()
return respectively gamma(1/n) and factorial(1/n), which are wrong results.

In this pull request, _eval_as_leading_term() functions are implemented
for both gamma and factorial functions.
This is in the direction of issue 17463.
Now the return values, when the arguments tend to infinity, are the
original functions. Probably the correct return value should be None, but
for compatibility conditions it is for the moment better the former one.

Incidentally, this PR corrects also the wrong results of issue 17431.