Move the code in clashing() in abc.py to the module level
This makes linters such as pyflakes not complain about undefined names in the
function. This way is preferable regardless because the doctests for the
function, which have been moved to the module docstring, can actually be run.
Previously, clashing was deleted at the end of the module, so it was
impossible for the doctester to run it.

See https://github.com/sympy/sympy/pull/17544#issuecomment-526373164.