Quaternion: Fixed the order of multiplication with a complex number
Fixed an issue that when a complex number on the left is multiplied by
a quaternion on the right, the multiplication is performed in the
reversed order.

Example before the fix:
-----------------------

In [1]: from sympy import Quaternion, I

In [2]: J = Quaternion(0, 0, 1, 0)

In [3]: K = Quaternion(0, 0, 0, 1)

In [4]: I * J == K
Out[4]: False

In [5]: I * J == -K
Out[5]: True

-----------------------

I also added a test to make sure that multiplying a complex number by
a quaternion is performed in the correct order. (And I corrected a
typo.)