Fix the see also links in the variations() docstring
The functions were removed from compatibility and are not just in the standard
library itertools module. See also doesn't support linking to external docs
directly, so we manually use the Sphinx seealso directive. We could also use
intersphinx, but that would make the build require internet to succeed, and
could potentially break other things that accidentally link to the Python
docs, so I haven't done that for now.