Make references in See Also sections work without a fully qualified name
As long as they available as sympy.name. In other words, writing 'function' in
a see also section is now the same as if you wrote '~.function' in a normal
Sphinx cross reference. This is only done for references that don't already
have dots.

This way, functions that are in the global SymPy namespace are not referenced
by their fully qualified name, which is preferred as the full module name of a
function that is in the global namespace should not be considered part of the
API (we could move it around without considering it to be an API break).

This fixes #17619.