Explicit error when lower bound is not given to the prime sieve.
sieve[:10] was failing with a TypeError in Python 3 and was returning
an empty list in Python 2.

We could allow sieve[:10] by defaulting to start by 1 in Python 3
only, but not in Python 2 as the start attribute of the slice is
receiving zero in this case, so we can't distinguish between
sieve[:10] and sive[0:10].

But as explicit is better, let's just disallow sieve[:10] and
sieve[0].

Also implemented __iter__ so iterating over the sieve don't
raise. Previously, iterating over a sieve was starting with the last
one (index 0) followed by the normal ones in normal order, which was
strange::

    >>> from sympy.ntheory import sieve
    >>> for i, prime in enumerate(sieve):
    ...     if i > 5:
    ...         break
    ...     print(prime)
    ...
    13
    2
    3
    5
    7
    11