Add parse_expr to the main __init__.py
It's basically been public API for a while, since it's the only way to access
the advanced parsing features like the input transformers.

I've not added the transformers themselves. They can still be accessed from
the sympy_parser submodule. It seems to me that adding them to __init__.py as
well would pollute the namespace, although I don't feel strongly about it.