tensor: Implemented tensorial and symbolic derivatives
Implemented `_eval_derivatives` for `TensAdd`, `TensMul`, and `Tensor`.
Further, added `_diff_wrt = True` and `is_scalar = True` to `Tensor`.

The implementation for `_eval_derivatives` are more or less taken from
`Mul` and `Add`. Derivatives with respect to tensors are implemented
in such a manner, that a covariant index in the derivative is
counted contravariant, and vice versa. Therefore, `diff(a(i), a(i))`
is a perfectly valid expression.

It is assumed that free indices (i.e. args[1] of `Tensor`) are always
in the correct order such that for `diff(a(i, j), a(k, l))` both args[1]
give (i, j) and (k, l), respectively.

Further, the implementation for tensorial derivates is -
besides fulfilling product and sum rule - given by (**):

* \partial a_i/\partial a_k = delta_i^k
* \partial a_i/\partial a^k = g_{ij} delta^j_k
* \partial a^i/\partial a^k = delta^i_k
* \partial a^i/\partial a_k = g^{ij} delta_j^k

Known issues and open problems:

* Does `is_scalar = True` lead to any inconsistencies?
* What about anticommuting/non-commuting tensors?
  Is there anything special to be considered?
* (**) What if there exists no metric? `diff(a(i), a(-k)) = 0`?
  Is this case even possible/considered in `tensor`?
* `sqrt(a(i)*a(-i))` cannot be derived by `a(i)` since `__div__`
  allows no tensor expressions, see (***).

Examples:

>>> Euclidean = TensorIndexType('Euclidean', dummy_name='E')
>>> a = tensor_heads('a', [Euclidean])
>>> b = tensor_heads('b', [Euclidean])
>>> i, j, k = tensor_indices('i j k', Euclidean)
>>> expr1 = a(i) + tau**2*b(i)
>>> expr1.diff(tau).doit()
2*tau*b(i)
>>> expr2 = a(i)*a(-i) + tau**2 + tau*b(i)*b(-i)
>>> expr2.diff(tau).doit()
2*tau + b(E_0)*b(-E_0)
>>> expr3 = myexpr2.diff(a(k))
>>> expr3.doit().contract_metric(i.tensor_index_type.metric).canon_bp()
2*KD(-k, E_0)*a(-E_0)
>>> expr4 = a(i)
>>> expr4.diff(a(i)).contract_metric(i.tensor_index_type.metric).\
    contract_delta(i.tensor_index_type.delta)
dim_E

TODO:

* doctests to be implemented
* tests to be implemented

Future nice to have features:

* Implement an abstract partial/covariant tensor derivative, e.g.:
  \partial_i (A_k*B_l) = (\partial_i A_k) B_l + A_k (\partial_i B_l)
  (Like `PD[][]` and `CD[][]` in `xTensor` [1])
* (***) Improve tensorial scalar functions support,
  see `xTensor` [1] `Scalar[]` head.

References:

[1] http://xact.es/xTensor