tensor: Introduced _eval_partial_derivative. Added class TensPartialDerivative
This commit takes the following points of discussion into account:

* `diff` should lead to coordinate invariant expressions
* this cannot be guaranteed for partial derivatives
* `_eval_derivative` should be renamed into `_eval_partial_derivative`
* a new class `TensPartialDerivative` should be introduced
* this could provide a method for handling abstract partial derivatives

Therefore, the method `_eval_derivative` was renamed to
`_eval_partial_derivative`. This lead to some new `if` statements
(for taking care, whether a `TensExpr` is differentiated or some other
entity), since one could not rely on the `diff` functionality of the super
classes very well.

Further the `TensPartialDerivative` class (a subclass of `TensExpr`) is
introduced, where the `doit` method calls `_eval_partial_derivative`.
The indexing of the unperformed abstract derivative remains to be discussed.