polys: Fixed the issue of expansion of irrational conjugates when calling cancel()
--> referring to issue #18205
>>> cancel((2+I)*(3-I))
>>> 7+I

>>> cancel((2+I)*(2-I))
>>> (2+I)*(2-I)

whereas cancel((2+I)*(3-I)) should return (2 + I)*(3 - I).
The cause of a expanded output was the Generator I in opt.gens,
which raised PolificationError in case of cancel((2+I)*(2-I)) as in this
case opt.gens was empty but in case of cancel((2+I)*(3-I)) opt.gens was not empty
it has I as a generator. Adding an if statement in poytools.py did the trick.

--> Added tests for same
--> Added dotprodsimp in sympy/simplify/init.py because it was causing Importerror after I pulled from master