core: Added new test case for verifying if replace method works with Equality expression
Added a new test case to check if the replace method exhibits correct behaviour with
Equality and in turn Relational. There was a bug in sympy 1.5.1 where the test case
mentioned below didn't work. Even though it was resolved due to a fix of some other
issue, this test case will still prove necessary for future versions to get the desired
results when replace and Equality/Relational are used in conjunction.

Example:

>> import sympy as sp
>> x, y, z = sp.symbols("x, y, z")
>> Xeq = sp.Eq(x, sp.pi * y**3 * (sp.Rational(1, 6) + z / 4 / y))
>> Xeq.replace(y**3, z)x
Eq(x, pi*z*(1/6 + z/(4*y)))