ntheory: Addition of chinese remainder theorem over cartesian product of vectors
In this commit, the changes introuduced in PR for issue #10949 has been added.
The issue comprised of implementing chinese remainder theorem over cartesian product
of vectors. So, the changes that were introuduced in that commit has been added in
this commit which includes implementation and test cases of the theorem.

Example:

>>> from sympy.ntheory.modular import crt_cartesian
>>> crt_cartesian([[3,5], [3,7]], [7, 11])
[3, 73, 47, 40]
>>> [(p%7, p%11) for p in [3, 40, 47, 73]]
[(3, 3), (5, 7), (5, 3), (3, 7)]