integrals: Minor changes for rebuild
Minor changes were made in this commit as the travis bot
failed unexpectedly in an unrelated test. Along with that,
the test_integrals.py  has uneven gaps between some of the
test functions and that was fixed in this commit as well.

integrals: Minor changes in the code

integrals: Added a comment for rebuild

Last time the build process stopped due to reasons not found in the
code so minor change was made for new build.

integrals: Added exception for a special case

When evaluating summation of integrals, if the integration variable
appears in the limits of a sum, then this ValueError is raised.