integrals: Added an exception for limits of a sum of integral
When the sum which is to be integrated has non-integer or non-finite
limits, a ValueError will be raised.

integrals: Case added for handling infinite summations in an integral

When a summation is infinite or there is a non-finite term in the
expansion of a summation, the integral of summation will not be
evaluated.

integrals: Minor update for rebuild

integrals: Updated the summation of integrations part

Updated to an older solution for the fix where the failing
test case was fixed to the newer definition of integrating.