Faster Determinant Solving by adding Triangular Matrix Check
Check for Triangular Matrix added in Determinant Solver Function
Earlier, for all matrices, either 'bareiss' or the specified
method was used for computing the determinant.
Now it is first checked whether it is a triangular matrix
and if it is, it allows determinant computation by just
product of diagonal elements

Similar _charpoly has been modified to check Triangular Matrix.
If it is found, diagonal elememts are eigen values also, and
they are used to compute the characteristic polynomial

Also updated the docstrings, added some more tests for
_charpoly and _det