Fix(Poly): Deprecate mxing Poly and Expr in binary operations
Add back the ability to mix Poly and Expr in binary operations like +,
-, and * even if the Expr can not be converted to a Poly. Convert the
Poly to Expr using as_expr so that the result evaluates to Expr. Emit a
SymPyDeprecationWarning when that happens. A previous commit removed
this functionality but this one adds it back with a deprecation warning.

The deprecated functionality is mixing Poly and Expr as in

    >>> from sympy import *
    >>> from sympy.abc import x
    >>> Poly(x, x) * exp(x)
    /Users/enojb/current/sympy/sympy/sympy/polys/polytools.py:70: SymPyDeprecationWarning:

    Mixing Poly with non-polynomial expressions in binary operations has
    been deprecated since SymPy 1.6. Use the as_expr or as_poly method to
    convert types instead. See https://github.com/sympy/sympy/issues/18613
    for more info.

      SymPyDeprecationWarning(
    x*exp(x)

BREAKING CHANGE: Mixing Poly and Expr in binary operations like + is now
deprecated. Use as_expr or as_poly to convert between Poly and Expr
before using binary operations.