Fix(Poly): Deprecate using Poly with integrate/Integral
Emit a deprecation warning when a Poly instances is passed as the
integrand to integrate or Integral. For example

  >>> from sympy import *
  >>> from sympy.abc import x
  >>> integrate(Poly(x, x))
  sympy/integrals/integrals.py:83: SymPyDeprecationWarning:

  Using integrate/Integral with Poly has been deprecated since SymPy
  1.6. Use the as_expr or integrate methods of Poly instead. See
  https://github.com/sympy/sympy/issues/18613 for more info.

    SymPyDeprecationWarning(
  Poly(1/2*x**2, x, domain='QQ')

Instead use the Poly.integrate method or as_expr to convert a Poly to
Expr.

BREAKING CHANGE: Deprecate using Poly instead of Expr as the integrand
argument to integrate or Integral. Use the integrate method or as_expr
instead.