Fix(old assumptions): Define infinite as not finite
Change the implication rule for infinite and finite in the old
assumptions from
  infinite -> !finite
to
  infinite == !finite

The effect of this is to add a new implication !finite->infinite which
means that is_infinite is the same as fuzzy_not(finite). Previously
inferring that an object had infinite=True would require a specific
_eval_is_infinite handler even if all of the logic was correctly
implemented in _eval_is_finite.