printing: Added parenthesis for substitution latex printing and revised tests
When printing expressions with substitutions, it may be hard to interpret
where the substitution begins, and where it ends. This commit adds parenthesis
to the latex output around the substitution in the expression, making the
output easier to interpret.

Example:
In[1]: from sympy import *
In[2]: from sympy.abc import x, y
In[3]: expr = 3*Subs(-x+y, (x,),(1,))
In[4]: print(latex(expr))
3 \left. \left(- x + y\right) \right|_{\substack{ x=1 }}
In[5]: pprint(expr)
3⋅(-x + y)│
          │x=1