fix(solvers): solve_linear_system calls linsolve
This makes solve_linear_system a wrapper for linsolve. Since linsolve
uses Matrix.rref which has seen recent improvements this can fix some
bugs and performance issues related to using solve_linear_system. This
also reduces the number of duplicate rref algorithms within the
codebase (GH-4949).

One test that has conditionally existent solutions is made an XFAIL. The
test is flakey and represents a behaviour of solve that does not hold in
general.

A couple of issues (GH-6420, GH-18531) with solve_linear_system are
fixed with tests added.