Handle the case where a character is assigned to an integer in C parser
C parser code could not handle the situation, where a variable which is declared as integer has been assigned with a character literal. For this, the function transform_character_literal() has been completed so as to handle character literals to be used in this case. Then, in transform_var_decl() function, after getting the node value from previous function, this case is handled in a separate elif block along with the comment giving its explanation. Also, corresponding test is added.