perf(solvers): make linear_eq_to_matrix faster
Improve performance of linear_eq_to_matrix for large sparse systems of
linear equations. A number of places in linear_eq_to_matrix had
quadratic behaviour in the number of symbols. This improves performance
where the number of symbols in each equation is small compared to the
total number of equations in the system.