perf(solvers): use connected_components in solve
Split systems of equations into connected components to be solved
independently so e.g. given the system

    x + y = 1
    y - x = 2
    z + t = 3
    t - z = 4

The first two equations can be solved for x and y and the second two can
be solved for t and z and the results combined. This leads to large
speed ups for many large sparse systems of either linear or polynomial
(or mixed) equations.

A number of places in solve are optimised for the case of solving a
large number of equations or returning a large number of solutions
because of bottlenecks thaht show up when _solve_system is made much
more efficient for these cases.