FIX: ensure autoscale_view gets called
In mpl <=3.1 the view limits were implicitly and unconditionally
autoscaled as part of setting the scale.  It was by co-incident that
by unconditionally setting the x and y scales when plotting also
triggered the auto-scale logic.  In mpl 3.2+ the view limits are
only autoscaled if changing the scale would could change the
acceptable limits.  As in the default case sympy is changing from
linear -> linear, this autoscaling was no longer applied.

This adds an explicit call to autoscaling to sympy will ensure it
behaves as desired independent of Matplotlib version.

Closes #19051