sanitize Eq/Ne args; Eq(True,1) -> True
Python treats True/False like 1/0 but in SymPy
we have a separation of S.true and S.One and they
are not equivalent.

In Eq/Ne when one arg is a non-symbol Boolean and
the other is 1 or 0 the 1/0 is interpreted and
converted to S.true/S.false. This means that
Eq(True, 1) is now True because the 1 will be
converted to S.true.

(It would be possible to only do the conversion
if the Boolean was less trivial, like ~x or
And(x, y), etc...