Define is_meromorphic for functions.
Most functions defined in SymPy are meromorphic with a few essential
singularities or branch points. A private attribute, ``_singularities``,
is added to function classes. For meromorphic functions, it is a finite
iterator that contains the essential singularities and branch points.
It may also be ``True``, ``False`` or ``None``, in which case it will
be the return value of ``is_meromorphic``.