tighten conditions on gamma rewrite
It is tempting to return an unconditional

gamma(n + 1)/(gamma(k + 1)*gamma(n - k + 1))

The problem is that when an arg in the denominator is zoo
then there is the immediate possibility of getting zoo/(zoo*zoo)
when k = -1, n = -2 is substituted, and this would give nan instead
of 0. If only a single argument of the 2 denominators were zoo then
the correct value of 0 would be returned. The problem is that if
the denominator is unevaluated and the expression is multiplied by
a gamma with the same argument the two arguments would silently
cancel:

    >>> gamma(n + 1)/(gamma(k + 1)*gamma(n - k + 1))*gamma(k + 1)
    gamma(n + 1)/gamma(n - k + 1)
    >>> _.subs(k,-2).subs(n,0)
    1/2

This value of 1/2 is the correct limiting value but if the terms
were not cancelled, a nan would be the result since 0*zoo is nan.