Remove expand_mul call in cholesky and LDLdecompisition
Both cholesky and LDLdecompisition used expand_mul to simplify some
calculations but this can lead to significant slowdowns in some cases.
This commit removes the calls to expand_mul from each of these
functions. A few tests needed to be changed because the output is no
longer simplified to the same extent.

See https://github.com/sympy/sympy/issues/19200