fix(testing): Set the warning filter around each test
When running tests/doctests with sympy's own test runners a warning
filter is used to turn DeprecationWarning/SymPyDeprecationWarning into
an error. The filter was only set one at the start of running the tests.
This means that if the filter is modified by one test then it will still
be modified when subsequent tests run.

This commit adds a raise_on_deprecated context manager in runtests.py
and uses it to set the warning filter in the test runners. The filter is
set immediately around running the test and is reset before each
individual test. This has picked up on a couple of failures that will
need to be fixed in subsequent commits.