19326: ordering subs dict
To make substitution canonical, unordered
replacements must be ordered in some way.
Although more complex topological sorting
is possible, the current method is to put
more complex patterns first and for those
that have the same complexity, order them
by using the default_sort_key.

This commit corrects the original change
which put more complex patterns first but
did not account for the keeping order amongst
the patterns that had the same complexity.

The _nodes function had to be modified to put
higher order derivatives ahead of those with
lower order.