Avoid repeated checks for superscripts in parenthesize_super
The behavior before was:

1. Check for a superscript. If so, call `parenthesize_super`, which...
2. Checks again for a superscript. If so call `embed_super`, which...
3. Checks again for a superscript.

The check in 1 can be immediately removed.
`embed_super` has exactly one caller, and it's easier just to inline it. This allows the check in 3 to be removed.

Making this change means subclasses can more easily override parenthesize_super, eg to also parenthesize subscripts if they so wish.