solveset: improve main trig solver (_solve_trig1)
Fixes #18315 (and others).

This commit adds some pre-processing to _solve_trig1. The motivation
of this processing is to fix the change of variable exp(I*x) -> y
for equations where some coefficients are non-integers. The actual code
is an enhanced version of the one found in _solve_trig2.

Before:
_solve_trig1(sin(x/2), x, Complexes)
_solve_trig1(sin(pi*x), x, Complexes)
_solve_trig1(sin(sqrt(2)*x), x, Complexes)
_solve_trig1(sin(a*x), x, Complexes)
all result in ConditionSets due to a failing 'subs' and execution passes
on to _solve_trig2 which already has this pre-processing. (Which is
incomplete, so that uncaught exceptions are raised...)
Also: _solve_trig1(sin(5*x)+cos(10*x), x, Complexes) results in an equation
of degree 20(!) even though a simple change of variable could take this
down to 4.

After:
The equation types above are handled fine.