sets: This implements the indicator functions for set theory.
Implements the indicator function (or characteristic function) which maps a number x
to the range {0, 1} for a set A; the function returns 1 if x is in A, and 0 if x is
not in A. This implementation works for defined numbers, and sets as well as symbolic
numbers and sets. The function, indicator, also includes the inverse of the indicator
function, the Mobius function (which has already been added for number theory). This
also adds relevant tests for the indicator function.

Examples:

In [1]: from sympy.sets import Union, Interval, FiniteSet, indicator
In [2]: set = FiniteSet(1, 2, 3)
In [3]: indicator(0, set)
Out [3]: 0
In [4]: indicator(1, set)
Out [4]: 1

References:
https://en.wikipedia.org/wiki/Indicator_function
https://mathworld.wolfram.com/CharacteristicFunction.html