Modified order.py
References to other Issues or PRs
Fixes #19545

Brief description of what is fixed or changed

The code `Order(1/n - 3/(3*n +2), (n, oo))` returns `1/n`, which
is not a sharp result.
As a consequence `Sum(1/n - 3/(3*n +2), (n, 1, oo)).is_convergent()`
gives the wrong result `False`.
In the lines 208-214 of `order.py`, if the number of variables are >= 2,
then the expression is expanded and is handled in `expr.is_Add` case.
However in the case of a single variable,
the alghorithm for the other case seems better.
Factorizing the expression forces to use the better alghoritm in the case
of a single variable.

See also PR#18340

On branch issue-19545
Changes to be committed:
    modified:   sympy/concrete/tests/test_sums_products.py
    modified:   sympy/series/order.py
    modified:   sympy/series/tests/test_order.py