Remove weird __str__ overloads with extra arguments
`__str__` shouldn't be receiving a sympy printer, that's what _ sympystr is for.
These classes will inherit appropriate `__str__` and `__repr__` implemented in terms of it from their base classes.

This also removes the custom `_sympyrepr` - the default supplied by sympy is usually sufficient.

Note that for now, this preserves the old behavior of ignoring all the settings of the passed in printer.
Fixing this can come in a later PR, now that it's obvious that it's happening.