Add support for dvipdfm[x] in printing.preview
On a windows miktex installation, there is no `dvipdf`, but there is `dvipdfmx`.
A cursory google suggests the latter is a better alternative in almost all regards.

Since the arguments taken by the `m` variants are different to `dvipdf`, the `commandend` dictionary is tweaked to be keyed by the actual executable name, not the format.