Ensure EvaluateFalseTransformer flattens Mul inside Sub
When transforming the tree, the code replaces subtraction BinOps with
calls to SymPy's Mul class. In doing so it provides the original right
argument of the BinOp as the right argument of the new Mul.
If the `right` argument was itself a Mul, you are left with a nested Mul
situation that flatten() was designed to fix. It is not flattened by the
existing denesting at the end of the function because the overall node is
now an Add and so the flattening only affects nested addition.

To see the bug, try for example:
> import sympy
> from sympy.parsing.sympy_parser import parse_expr
> sympy.srepr(parse_expr("1-x*y", evaluate=False))
"Add(Integer(1), Mul(Integer(-1), Mul(Symbol('x'), Symbol('y'))))"

where clearly the second Mul could easily have been flattened into the
first, without changing the meaning or evaluating anything, to form:
"Add(Integer(1), Mul(Integer(-1), Symbol('x'), Symbol('y')))"
This change leads to the expected behaviour, but may affect cases where the
tree is expected to be in a specific format.