feat(solvers): split systems into connected components
Don't recompute the split after recursively calling solve. Since solve
computes the connected components of the system of equations and the
pass each back to solve recursively the connected components get
recomputed in the second call. This makes sure that the connected
components are not recomputed in the recursive call.