Do not use SystemError
The docs for system error are clear:
```
>>> print(SystemError.__doc__)
Internal error in the Python interpreter.

Please report this to the Python maintainer, along with the traceback,
the Python version, and the hardware/OS platform and version.
```
None of these errors justify contacting a python maintainer.

`OSError` was probably intended, although a few of these make more sense as `ValueError`.