Add Map.apply
This method is an API for map and its applied class.
Changed the classes in operator.py, add.py and mul.py to use this method. This
change made AppliedBinaryOperator needless, so it is deleted.
Also, relation between InverseElement and ExponentElement is modified. Now,
if the base operator is associative, ExponentElement is used instead of
InverseElement.