interactive: Fix a regression in ipython pretty-printing
If __repr__ is defined explicitly, then the pretty-printer connected by `init_printing` does not have high enough precedence to take effect.
The `__repr__` in this file is identical to the inherited one anyway, so there is no reason to define it.

This regression was introduced with the new `printing.defaults.Printable` class.