Remove sorting from BasicMeta
This breaks a lot of things (sympy itself can't even be imported). We will
need to go through the codebase and fix everything that relies on this
behavior. It will not be simple.

I have left the BasicMeta metaclass intact as an empty class because there is
some code that uses it to check if a class is a Basic class. We should
probably rename ManagedProperties to BasicMeta, but for now, I don't want to
break compatibility.