Fix 'from sympy.printing import ccode' after running 'import sympy.printing.ccode'
If ccode.py is not yet imported, then sympy.printing.ccode will be the
function, as imported in sympy/printing/__init__.py from
sympy/printing/codeprinter.py. But once sympy.printing.ccode is imported,
Python replaces it with the submodule, which breaks the above import.

To fix this, we have to make sure that the submodules ccode.py and fcode.py
are imported before the functions in sympy/printing/__init__.py. This is
somewhat unfortunate, as we would otherwise not even need to import those
submodules at all at top-level import time.

The only other alternative way to fix this that I can think of is to disallow
'from sympy.printing import ccode' and require 'from sympy import ccode'.
However, the former form already appears in several docstring examples, and it
is generally the sort of thing that people do in actual code, so it would be
very confusing to let this behavior happen.